+++ src/core/vt_database_evaluate.c
@@ -17,10 +17,10 @@ int _vt_database_evaluate_nrmse(VT_DATABASE* database_ptr, uint32_t* fallcurvear
     float min = 65355.00;
     int index;
 
-    for (uint32_t i = 0; i < (database_ptr->_vt_total_fingerprints); i++)
+    for (uint8_t i = 0; i < (database_ptr->_vt_total_fingerprints); i++)
         nrmse[i] = _vt_fingerprint_evaluate_nrmse(&(database_ptr->_vt_fingerprintdb[i][2]), fallcurvearray, 100);
 
-    for (uint32_t k = 0; k < (database_ptr->_vt_total_fingerprints); k++)
+    for (uint8_t k = 0; k < (database_ptr->_vt_total_fingerprints); k++)
         if (nrmse[k] < min)
         {
             min = nrmse[k], index = k;
@@ -74,7 +74,7 @@ int _vt_database_evaluate_pearson_falltime(VT_DATABASE* database_ptr, int fall_t
 
 static int _vt_database_falltime_nearestindex_search(VT_DATABASE* database_ptr, uint32_t fall_time)
 {
-    uint32_t i;
+    uint8_t i;
 
     for (i = 0; (i < database_ptr->_vt_total_falltime) && (database_ptr->_vt_falltimedb[i][1] <= fall_time); i++)
         ;
